refactor: move app folder to views
