+++ src/renderer/views/app/components/App/style.ts
