+++ src/renderer/views/app/components/HorizontalScrollbar/index.tsx
