+++ src/renderer/views/app/components/HorizontalScrollbar/style.ts
