+++ src/renderer/views/app/components/Tabbar/style.ts
