+++ src/renderer/views/app/components/Tabs/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { observer } from 'mobx-react-lite';
 
-import store from '~/renderer/app/store';
+import store from '~/renderer/views/app/store';
 import Tab from '../Tab';
 
 export const Tabs = observer(() => {
