+++ src/renderer/views/app/constants/design.ts
