+++ src/renderer/views/app/constants/icons.ts
