+++ src/renderer/views/app/constants/tabs.ts
