+++ src/renderer/views/app/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
 import { App } from './components/App';
-import { fonts } from '../constants';
+import { fonts } from '../../constants';
 import { ipcRenderer } from 'electron';
 
 ipcRenderer.setMaxListeners(0);
