+++ src/renderer/views/app/utils/colors.ts
